Moved all tests to proper folders. All highest directories are lowercase now.
