+++ homepairs/homepairsUI/babel.config.js
@@ -27,14 +27,14 @@ module.exports = function(api) {
                         'homepairs-strings': './res/strings.tsx',
                         'homepairs-colors': './res/colors.tsx',
                         'homepairs-types': './src/state/types.ts',
-                        'homepairs-elements': './src/Elements/index.tsx',
+                        'homepairs-elements': './src/elements/index.tsx',
                         'homepairs-redux-actions': './src/state/actions.ts',
-                        'homepairs-pages': './src/Screens/index.tsx',
-                        'homepairs-modals': './src/Modals/index.tsx',
+                        'homepairs-pages': './src/screens/index.tsx',
+                        'homepairs-modals': './src/modals/index.tsx',
                         'homepairs-base-styles': './res/Styles/base.ts',
                         'homepairs-components': './src/Components/index.tsx',
                         'homepairs-utilities': './src/utility/index.tsx',
-                        'homepairs-routes' : './src/Routes/index.tsx',
+                        'homepairs-routes' : './src/routes/index.tsx',
                         'homepairs-endpoints' : './src/endpoints/index.ts',
                         'homepairs-test' : './tests/fixtures/index',
                     },
