+++ homepairs/homepairsUI/src/endpoints/__test__/GenerateTenantAccount.test.tsx
@@ -2,10 +2,9 @@ import { AccountActions, PropertyListActions } from 'homepairs-redux-actions';
 import { AccountTypes, AccountStateAction, Account, FetchPropertyAndPropertyManagerAction, SetAccountAuthenticationStateAction} from 'homepairs-types';
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import { navigationPages} from 'src/Routes/RouteConstants';
-import { HOMEPAIRS_REGISTER_TENANT_ENDPOINT, generateAccountForTenant } from 'homepairs-endpoints';
-import { propertyManagerMock1 } from '../../fixtures/StoreFixture';
-import {prepareNavigationSwitchMock} from '../../fixtures/DummyComponents';
+import { navigationPages} from 'homepairs-routes';
+import { propertyManagerMock1 , prepareNavigationSwitchMock} from 'homepairs-test';
+import { HOMEPAIRS_REGISTER_TENANT_ENDPOINT, generateAccountForTenant } from '../index';
 
 
 const {FETCH_PROFILE} = AccountActions.FETCH_PROFILE_ACTION_TYPES;
