+++ homepairs/homepairsUI/src/endpoints/__test__/PostUpdatedProperty.test.tsx
@@ -1,74 +1,75 @@
 import { PropertyListActions } from 'homepairs-redux-actions';
 import {
     Property,
-    AddNewPropertyState,
-    AddPropertyAction,
+    EditPropertyState,
+    UpdatePropertyAction,
 } from 'homepairs-types';
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import {HOMEPAIRS_PROPERTY_ENDPOINT, postNewProperty } from 'homepairs-endpoints';
-
-import { prepareNavigationMock, displayErrorMock } from 'tests/fixtures/DummyComponents';
-import { propertyManagerMock1 } from '../../fixtures/StoreFixture';
+import { prepareNavigationMock, displayErrorMock, propertyManagerMock1 } from 'homepairs-test';
+import { HOMEPAIRS_PROPERTY_ENDPOINT, postUpdatedProperty } from '../index';
 
 const URL = HOMEPAIRS_PROPERTY_ENDPOINT;
-const {PROPERTY_LIST_ACTION_TYPES } = PropertyListActions;
-const [mockStackNavigator, navigationStackSpyFunction] = prepareNavigationMock();
+const { PROPERTY_LIST_ACTION_TYPES } = PropertyListActions;
+const [mockNavigator, mockNavigationFunction] = prepareNavigationMock();
 
-const newPropertyTest: Property = {
-    address: 'Cool Aid Drive, HoodRat City, GA',
+const prevProperty: Property = {
+    address: 'Cool Aid Road',
+    tenants: 4,
+    bedrooms: 3,
+    bathrooms: 3,
+};
+const updatedPropertyTest: Property = {
+    address: 'Cool Aid Drive',
     tenants: 3,
     bedrooms: 3,
     bathrooms: 3,
 };
-const newPropertyCredentials: AddNewPropertyState = {
+const updatedPropertyCredentials: EditPropertyState = {
     email: 'jackylynne@gmail.com',
+    index: 0,
+    oldProp: prevProperty,
     roopairsToken: '1B88484KDKSLIEK058UDDD',
 };
 
 // To test nested actions with an async REST request, we need to call an await on the
 // dispatch methods in order to get an updated state.
-describe('postNewProperty Action', () => {
+describe('postUpdatedProperty Action', () => {
     const mock = new MockAdapter(axios);
 
     describe('Tests Action when response returned is success', () => {
-        const setInitialStateSpy = jest.fn();
 
         beforeEach(() => {
             propertyManagerMock1.clearActions();
-            navigationStackSpyFunction.mockClear();
-            setInitialStateSpy.mockClear();
-            displayErrorMock.mockClear();
+            mockNavigationFunction.mockClear();
         });
 
         it('Test the dispatch and callback methods', async () => {
             const data = {
                 status: 'success',
             };
-            const expectedResult: AddPropertyAction = {
-                type: PROPERTY_LIST_ACTION_TYPES.ADD_PROPERTY,
+            const expectedResult: UpdatePropertyAction = {
+                type: PROPERTY_LIST_ACTION_TYPES.UPDATE_PROPERTY,
+                index: 0,
                 userData: {
-                    address: 'Cool Aid Drive, HoodRat City, GA',
+                    address: 'Cool Aid Drive',
                     tenants: 3,
                     bedrooms: 3,
                     bathrooms: 3,
-                    propId: undefined,
                 },
             };
 
-            mock.onPost(URL).reply(200, data);
+            mock.onPut(URL).reply(200, data);
             await propertyManagerMock1
                 .dispatch(
-                    postNewProperty(
-                        newPropertyTest,
-                        newPropertyCredentials,
-                        setInitialStateSpy,
+                    postUpdatedProperty(
+                        updatedPropertyTest,
+                        updatedPropertyCredentials,
                         displayErrorMock,
-                        mockStackNavigator,
+                        mockNavigator,
                     ),
                 )
                 .then(() => {
-                    expect(setInitialStateSpy).toHaveBeenCalledTimes(1);
                     const actionResults = propertyManagerMock1.getActions();
                     expect(actionResults).toHaveLength(1);
                     expect(actionResults[0]).toStrictEqual(expectedResult);
@@ -77,32 +78,26 @@ describe('postNewProperty Action', () => {
     });
 
     describe('Tests Action when on failure', () => {
-        const setInitialStateSpy = jest.fn();
-
         beforeEach(() => {
             propertyManagerMock1.clearActions();
-            navigationStackSpyFunction.mockClear();
-            setInitialStateSpy.mockClear();
-            displayErrorMock.mockClear();
+            mockNavigationFunction.mockClear();
         });
 
         it('Test when failure to get response', async () => {
-            mock.onPost(URL).reply(404);
+            mock.onPut(URL).reply(500);
             await propertyManagerMock1
                 .dispatch(
-                    postNewProperty(
-                        newPropertyTest,
-                        newPropertyCredentials,
-                        setInitialStateSpy,
+                    postUpdatedProperty(
+                        updatedPropertyTest,
+                        updatedPropertyCredentials,
                         displayErrorMock,
-                        mockStackNavigator,
+                        mockNavigator,
                     ),
                 )
                 .then(() => {
-                    expect(setInitialStateSpy).toHaveBeenCalledTimes(0);
+                    expect(mockNavigationFunction).toHaveBeenCalledTimes(0);
                     const actionResults = propertyManagerMock1.getActions();
                     expect(actionResults).toHaveLength(0);
-                    expect(navigationStackSpyFunction).toHaveBeenCalledTimes(0);                    
                 });
         });
 
@@ -110,22 +105,20 @@ describe('postNewProperty Action', () => {
             const data = {
                 status: 'failure',
             };
-            mock.onPost(URL).reply(200, data);
+            mock.onPut(URL).reply(200, data);
             await propertyManagerMock1
                 .dispatch(
-                    postNewProperty(
-                        newPropertyTest,
-                        newPropertyCredentials,
-                        setInitialStateSpy,
+                    postUpdatedProperty(
+                        updatedPropertyTest,
+                        updatedPropertyCredentials,
                         displayErrorMock,
-                        mockStackNavigator,
+                        mockNavigator,
                     ),
                 )
                 .then(() => {
-                    expect(setInitialStateSpy).toHaveBeenCalledTimes(0);
+                    expect(mockNavigationFunction).toHaveBeenCalledTimes(0);
                     const actionResults = propertyManagerMock1.getActions();
                     expect(actionResults).toHaveLength(0);
-                    expect(navigationStackSpyFunction).toHaveBeenCalledTimes(0);                    
                 });
         });
     });
