+++ homepairs/homepairsUI/src/utility/__test__/ParameterChecker.test.tsx
@@ -1,4 +1,4 @@
-import {isNullOrUndefined} from 'homepairs-utilities';
+import {isNullOrUndefined} from '../ParameterChecker';
 
 type optionalProps = {
     test1?: any
