+++ homepairs/homepairsUI/tsconfig.jest.json
@@ -11,14 +11,14 @@
         "homepairs-strings": "./res/strings.tsx",
         "homepairs-colors": "./res/colors.tsx",
         "homepairs-types": "./src/state/types.ts",
-        "homepairs-elements": "./src/Elements/index.tsx",
+        "homepairs-elements": "./src/elements/index.tsx",
         "homepairs-redux-actions": "./src/state/actions.ts",
-        "homepairs-pages": "./src/Screens/index.ts",
+        "homepairs-pages": "./src/screens/index.ts",
         "homepairs-base-styles": "./res/Styles/base.ts",
-        "homepairs-components": "./src/Screens/Components/index.tsx",
-        "homepairs-modals": "./src/Modals/index.tsx",
+        "homepairs-components": "./src/components/index.tsx",
+        "homepairs-modals": "./src/modals/index.tsx",
         "homepairs-utilities": "./src/utility/index.tsx",
-        "homepairs-routes" : "./src/Routes/index.tsx",
+        "homepairs-routes" : "./src/routes/index.tsx",
         "homepairs-test" : "./tests/fixtures/index",
     },
 }
