+++ homepairs/homepairsUI/tsconfig.json
@@ -21,14 +21,14 @@
             "homepairs-types": ["./src/state/types.ts"],
             "homepairs-strings": ["./res/strings.tsx"],
             "homepairs-colors": ["./res/colors.tsx"],
-            "homepairs-elements": ["./src/Elements/index.tsx"],
+            "homepairs-elements": ["./src/elements/index.tsx"],
             "homepairs-redux-actions": ["./src/state/actions.ts"],
-            "homepairs-pages": ["./src/Screens/index.tsx"],
-            "homepairs-modals": ["./src/Modals/index.tsx"],
+            "homepairs-pages": ["./src/screens/index.tsx"],
+            "homepairs-modals": ["./src/modals/index.tsx"],
             "homepairs-base-styles": ["./res/Styles/base.ts"],
             "homepairs-utilities": ["./src/utility/index.tsx"],
-            "homepairs-components": ["./src/Components/index.tsx"],
-            "homepairs-routes" : ["./src/Routes/index.tsx"],
+            "homepairs-components": ["./src/components/index.tsx"],
+            "homepairs-routes" : ["./src/routes/index.tsx"],
             "homepairs-endpoints" : ["./src/endpoints/index.ts"],
             "homepairs-test" : ["./tests/fixtures/index"],
         }
